EPG refactoring by SimonC
Split function into separate files
Simplify button press event handling
Move zapping code into mix in class
Enhance EPG preview mode to restart playback
Some Bug fixes
Fix multi EPG to remember position when adding timers
Use EPG startRef consistently
Allow more EPG history to be kept
Delete old EPG files- apart from what is required for other plugins
Fix disappearing selection bug
Stop IMDb opening if no event selected
--Open single EPG from Graph EPG if no event selected on current service
