+++ lib/python/Components/Epg/EpgBouquetList.py
@@ -1,0 +1,269 @@
+import skin
+from time import localtime, time, strftime
+
+from enigma import eEPGCache, eListbox, eListboxPythonMultiContent, loadPNG, gFont, getDesktop, eRect, eSize, RT_HALIGN_LEFT, RT_HALIGN_RIGHT, RT_HALIGN_CENTER, RT_VALIGN_CENTER, RT_VALIGN_TOP, RT_WRAP, BT_SCALE, BT_KEEP_ASPECT_RATIO, BT_ALIGN_CENTER
+
+from Components.GUIComponent import GUIComponent
+from Components.MultiContent import MultiContentEntryText, MultiContentEntryPixmapAlphaBlend, MultiContentEntryPixmapAlphaTest
+from skin import parseColor, parseFont
+from Tools.Alternatives import CompareWithAlternatives
+from Components.config import config
+from Tools.Directories import resolveFilename, SCOPE_ACTIVE_SKIN
+
+class EPGBouquetList(GUIComponent):
+	def __init__(self, graphic=False):
+		GUIComponent.__init__(self)
+		self.graphic = graphic
+		self.l = eListboxPythonMultiContent()
+		self.l.setBuildFunc(self.buildEntry)
+
+		self.onSelChanged = [ ]
+
+		self.foreColor = 0xffffff
+		self.foreColorSelected = 0xffffff
+		self.backColor = 0x2D455E
+		self.backColorSelected = 0xd69600
+
+		self.borderColor = 0xC0C0C0
+		self.BorderWidth = 1
+
+		self.othPix = None
+		self.selPix = None
+		self.graphicsloaded = False
+
+		self.bouquetFontName = "Regular"
+		self.bouquetFontSize = 20
+
+		self.itemHeight = 31
+		self.listHeight = None
+		self.listWidth = None
+
+		self.bouquetNamePadding = 3
+		self.bouquetNameAlign = 'left'
+		self.bouquetNameWrap = 'no'
+
+	def applySkin(self, desktop, screen):
+		if self.skinAttributes is not None:
+			attribs = [ ]
+			for (attrib, value) in self.skinAttributes:
+				if attrib == "font":
+					font = parseFont(value, ((1,1),(1,1)) )
+					self.bouquetFontName = font.family
+					self.bouquetFontSize = font.pointSize
+				elif attrib == "foregroundColor":
+					self.foreColor = parseColor(value).argb()
+				elif attrib == "backgroundColor":
+					self.backColor = parseColor(value).argb()
+				elif attrib == "foregroundColorSelected":
+					self.foreColorSelected = parseColor(value).argb()
+				elif attrib == "backgroundColorSelected":
+					self.backColorSelected = parseColor(value).argb()
+				elif attrib == "borderColor":
+					self.borderColor = parseColor(value).argb()
+				elif attrib == "borderWidth":
+					self.BorderWidth = int(value)
+				elif attrib == "itemHeight":
+					self.itemHeight = int(value)
+				else:
+					attribs.append((attrib,value))
+			self.skinAttributes = attribs
+		rc = GUIComponent.applySkin(self, desktop, screen)
+		self.setBouquetFontsize()
+		self.listHeight = self.instance.size().height()
+		self.listWidth = self.instance.size().width()
+		self.l.setItemHeight(self.itemHeight)
+		return rc
+
+	GUI_WIDGET = eListbox
+
+	def getCurrentBouquet(self):
+		return self.l.getCurrentSelection()[0]
+
+	def getCurrentBouquetService(self):
+		return self.l.getCurrentSelection()[1]
+
+	def setCurrentBouquet(self, CurrentBouquetService):
+		self.CurrentBouquetService = CurrentBouquetService
+
+	def selectionChanged(self):
+		for x in self.onSelChanged:
+			if x is not None:
+				x()
+
+	def getIndexFromService(self, serviceref):
+		if serviceref is not None:
+			for x in range(len(self.bouquetslist)):
+				if CompareWithAlternatives(self.bouquetslist[x][1], serviceref):
+					return x
+		return None
+
+	def moveToService(self, serviceref):
+		newIdx = self.getIndexFromService(serviceref)
+		if newIdx is None:
+			newIdx = 0
+		self.setCurrentIndex(newIdx)
+
+	def setCurrentIndex(self, index):
+		if self.instance is not None:
+			self.instance.moveSelectionTo(index)
+
+	def moveTo(self, dir):
+		if self.instance is not None:
+			self.instance.moveSelection(dir)
+
+	def setBouquetFontsize(self):
+		self.l.setFont(0, gFont(self.bouquetFontName, self.bouquetFontSize))
+
+	def postWidgetCreate(self, instance):
+		self.l.setSelectableFunc(True)
+		instance.setWrapAround(True)
+		instance.selectionChanged.get().append(self.selectionChanged)
+		instance.setContent(self.l)
+		# self.l.setSelectionClip(eRect(0,0,0,0), False)
+		self.setBouquetFontsize()
+
+	def preWidgetRemove(self, instance):
+		instance.selectionChanged.get().append(self.selectionChanged)
+		instance.setContent(None)
+
+	def selectionEnabled(self, enabled):
+		if self.instance is not None:
+			self.instance.setSelectionEnable(enabled)
+
+	def recalcEntrySize(self):
+		esize = self.l.getItemSize()
+		width = esize.width()
+		height = esize.height()
+		self.bouquet_rect = eRect(0, 0, width, height)
+
+	def getBouquetRect(self):
+		rc = self.bouquet_rect
+		return eRect( rc.left() + (self.instance and self.instance.position().x() or 0), rc.top(), rc.width(), rc.height() )
+
+	def buildEntry(self, name, func):
+		r1 = self.bouquet_rect
+		left = r1.left()
+		top = r1.top()
+		# width = (len(name)+5)*8
+		width = r1.width()
+		height = r1.height()
+		selected = self.CurrentBouquetService == func
+
+		if self.bouquetNameAlign.lower() == 'left':
+			if self.bouquetNameWrap.lower() == 'yes':
+				alignnment = RT_HALIGN_LEFT | RT_VALIGN_CENTER | RT_WRAP
+			else:
+				alignnment = RT_HALIGN_LEFT | RT_VALIGN_CENTER
+		else:
+			if self.bouquetNameWrap.lower() == 'yes':
+				alignnment = RT_HALIGN_CENTER | RT_VALIGN_CENTER | RT_WRAP
+			else:
+				alignnment = RT_HALIGN_CENTER | RT_VALIGN_CENTER
+
+		res = [ None ]
+
+		if selected:
+			if self.graphic:
+				borderTopPix = self.borderSelectedTopPix
+				borderLeftPix = self.borderSelectedLeftPix
+				borderBottomPix = self.borderSelectedBottomPix
+				borderRightPix = self.borderSelectedRightPix
+			foreColor = self.foreColor
+			backColor = self.backColor
+			foreColorSel = self.foreColorSelected
+			backColorSel = self.backColorSelected
+			bgpng = self.selPix
+			if bgpng is not None and self.graphic:
+				backColor = None
+				backColorSel = None
+		else:
+			if self.graphic:
+				borderTopPix = self.borderTopPix
+				borderLeftPix = self.borderLeftPix
+				borderBottomPix = self.borderBottomPix
+				borderRightPix = self.borderRightPix
+			backColor = self.backColor
+			foreColor = self.foreColor
+			foreColorSel = self.foreColorSelected
+			backColorSel = self.backColorSelected
+			bgpng = self.othPix
+			if bgpng is not None and self.graphic:
+				backColor = None
+				backColorSel = None
+
+		# box background
+		if bgpng is not None and self.graphic:
+			res.append(MultiContentEntryPixmapAlphaTest(
+				pos = (left + self.BorderWidth, top + self.BorderWidth),
+				size = (width - 2 * self.BorderWidth, height - 2 * self.BorderWidth),
+				png = bgpng,
+				flags = BT_SCALE))
+		else:
+			res.append(MultiContentEntryText(
+				pos = (left , top), size = (width, height),
+				font = 0, flags = RT_HALIGN_LEFT | RT_VALIGN_CENTER,
+				text = "", color = None, color_sel = None,
+				backcolor = backColor, backcolor_sel = backColorSel,
+				border_width = self.BorderWidth, border_color = self.borderColor))
+
+		evX = left + self.BorderWidth + self.bouquetNamePadding
+		evY = top + self.BorderWidth
+		evW = width - 2 * (self.BorderWidth + self.bouquetNamePadding)
+		evH = height - 2 * self.BorderWidth
+
+		res.append(MultiContentEntryText(
+			pos = (evX, evY), size = (evW, evH),
+			font = 0, flags = alignnment,
+			text = name,
+			color = foreColor, color_sel = foreColorSel,
+			backcolor = backColor, backcolor_sel = backColorSel))
+
+		# Borders
+		if self.graphic:
+			if borderTopPix is not None:
+				res.append(MultiContentEntryPixmapAlphaTest(
+						pos = (left, r1.top()),
+						size = (r1.width(), self.BorderWidth),
+						png = borderTopPix,
+						flags = BT_SCALE))
+			if borderBottomPix is not None:
+				res.append(MultiContentEntryPixmapAlphaTest(
+						pos = (left, r1.height()-self.BorderWidth),
+						size = (r1.width(), self.BorderWidth),
+						png = borderBottomPix,
+						flags = BT_SCALE))
+			if borderLeftPix is not None:
+				res.append(MultiContentEntryPixmapAlphaTest(
+						pos = (left, r1.top()),
+						size = (self.BorderWidth, r1.height()),
+						png = borderLeftPix,
+						flags = BT_SCALE))
+			if borderRightPix is not None:
+				res.append(MultiContentEntryPixmapAlphaTest(
+						pos = (r1.width()-self.BorderWidth, left),
+						size = (self.BorderWidth, r1.height()),
+						png = borderRightPix,
+						flags = BT_SCALE))
+
+		return res
+
+	def fillBouquetList(self, bouquets):
+		if self.graphic and not self.graphicsloaded:
+			self.othPix = loadPNG(resolveFilename(SCOPE_ACTIVE_SKIN, 'epg/OtherEvent.png'))
+			self.selPix = loadPNG(resolveFilename(SCOPE_ACTIVE_SKIN, 'epg/SelectedCurrentEvent.png'))
+
+			self.borderTopPix = loadPNG(resolveFilename(SCOPE_ACTIVE_SKIN, 'epg/BorderTop.png'))
+			self.borderBottomPix = loadPNG(resolveFilename(SCOPE_ACTIVE_SKIN, 'epg/BorderLeft.png'))
+			self.borderLeftPix = loadPNG(resolveFilename(SCOPE_ACTIVE_SKIN, 'epg/BorderBottom.png'))
+			self.borderRightPix = loadPNG(resolveFilename(SCOPE_ACTIVE_SKIN, 'epg/BorderRight.png'))
+			self.borderSelectedTopPix = loadPNG(resolveFilename(SCOPE_ACTIVE_SKIN, 'epg/SelectedBorderTop.png'))
+			self.borderSelectedLeftPix = loadPNG(resolveFilename(SCOPE_ACTIVE_SKIN, 'epg/SelectedBorderLeft.png'))
+			self.borderSelectedBottomPix = loadPNG(resolveFilename(SCOPE_ACTIVE_SKIN, 'epg/SelectedBorderBottom.png'))
+			self.borderSelectedRightPix = loadPNG(resolveFilename(SCOPE_ACTIVE_SKIN, 'epg/SelectedBorderRight.png'))
+
+			self.graphicsloaded = True
+		self.bouquetslist = bouquets
+		self.l.setList(self.bouquetslist)
+		self.recalcEntrySize()
+		self.selectionChanged()
+		self.CurrentBouquetService = self.getCurrentBouquetService()
