+++ lib/python/Components/Epg/Makefile.am
@@ -1,0 +1,3 @@
+installdir = $(pkglibdir)/python/Components/Epg
+
+install_PYTHON = *.py
