+++ lib/python/Components/Makefile.am
@@ -1,5 +1,5 @@
 installdir = $(pkglibdir)/python/Components
 
-SUBDIRS = Sources Converter Renderer
+SUBDIRS = Sources Converter Epg Renderer
 
 install_PYTHON = *.py
