+++ lib/python/Screens/Epg/Makefile.am
@@ -1,0 +1,4 @@
+installdir = $(pkglibdir)/python/Screens/Epg
+
+install_PYTHON = *.py
+
