+++ lib/python/Screens/Makefile.am
@@ -1,4 +1,6 @@
 installdir = $(pkglibdir)/python/Screens
 
+SUBDIRS = Epg
+
 install_PYTHON = *.py
